requirements.txt
mail-parser
pydantic
sqlalchemy
pytest
Pillow
python-magic
python-magic-bin; platform_system == 'Windows'
google-genai
pdf2image
python-dotenv
